Extract server, client, fetch-client into separate modules
