+++ zio-http-client/shared/src/main/scala/zio/http/Proxy.scala
