+++ zio-http-fetch-client/js/src/main/scala/zio/http/internal/FetchBodyBatched.scala
