+++ zio-http-server/shared/src/main/scala/zio/http/Driver.scala
@@ -27,8 +27,6 @@ trait Driver {
   def start(implicit trace: Trace): RIO[Scope, StartResult]
 
   def addApp[R](newRoutes: Routes[R, Response], env: ZEnvironment[R])(implicit trace: Trace): UIO[Unit]
-
-  def createClientDriver()(implicit trace: Trace): ZIO[Scope, Throwable, ClientDriver]
 }
 
 object Driver extends DriverPlatformSpecific {
